requirements.txt
gradio>=4.0.0
dspy-ai>=2.4.0
openai>=1.0.0
spacy>=3.7.0
firecrawl-py>=0.0.8
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
numpy>=1.24.0
ruff>=0.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl